[CHORE] : 변수명을 구체화하고 예외 핸들러의 위치 수정.
Close feature/이수영_#13
