+++ src/main/java/com/prgrms/be/intermark/domain/user/controller/UserController.java
@@ -1,19 +1,18 @@
 package com.prgrms.be.intermark.domain.user.controller;
 
-import com.prgrms.be.intermark.common.dto.ErrorResponse;
-import com.prgrms.be.intermark.common.dto.page.dto.PageResponseDTO;
+import com.prgrms.be.intermark.common.dto.page.PageResponseDTO;
 import com.prgrms.be.intermark.domain.user.User;
 import com.prgrms.be.intermark.domain.user.dto.UserInfoResponseDTO;
 import com.prgrms.be.intermark.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-import javax.persistence.EntityNotFoundException;
 import javax.websocket.server.PathParam;
-import java.time.LocalDateTime;
 
 @RestController
 @RequestMapping("/api/v1/users")
@@ -35,14 +34,5 @@ public class UserController {
             @PathParam(value = "size") int size) {
         PageResponseDTO<User, UserInfoResponseDTO> users = userService.findAllUser(PageRequest.of(page, size));
         return ResponseEntity.ok(users);
-    }
-
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    @ExceptionHandler(EntityNotFoundException.class)
-    public ErrorResponse entityNotFoundExceptionHandler(EntityNotFoundException exception) {
-        return ErrorResponse.of(
-                HttpStatus.NOT_FOUND,
-                exception.getMessage(),
-                LocalDateTime.now());
     }
 }
